pyproject.toml
[project]
name = "pharma-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=1.4.1",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.1.1",
    "pyjwt>=2.9.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.1.0",
    "email-validator",
    "python-dotenv>=1.0.0",
    "pandas>=2.3.3",
]
[tool.uv.workspace]
members = ["agents"]

